rename interface package, align structs. more comments
