+++ entry.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"slices"
 
-	"golift.io/goty/goatface"
+	"golift.io/goty/gotyface"
 )
 
 const (
@@ -39,7 +39,7 @@ type Config struct {
 	// GlobalOverrides are applied to all structs unless a type-specific override exists.
 	GlobalOverrides Override `json:"globalOverrides" toml:"global_overrides" xml:"global-override" yaml:"globalOverrides"`
 	// DocHandler is the handler for go/doc comments. Comments are off by default.
-	goatface.DocHandler `json:"-" toml:"-" xml:"-" yaml:"-"`
+	gotyface.DocHandler `json:"-" toml:"-" xml:"-" yaml:"-"`
 }
 
 // Overrides is a map of go types to their typescript override values.
@@ -47,19 +47,22 @@ type Overrides map[any]Override
 
 // Override is a struct that contains overrides for either a specific type or for all types (when global).
 type Override struct {
+	// Namer is a function that can be used to customize the typescript interface name.
+	// Use this to add a prefix, suffix or any custom name changes you wish.
+	Namer Namer `json:"-" toml:"-" xml:"-" yaml:"-"`
 	// Typescript type. ie. string, number, boolean, etc.
 	// This has no effect when set inside a global override; it's type specific.
 	Type string `json:"type" toml:"type" xml:"type" yaml:"type"`
 	// Typescript interface name. This does not work on field names.
 	// This has no effect when set inside a global override; it's type specific.
 	Name string `json:"name" toml:"name" xml:"name" yaml:"name"`
-	// Setting optional to true will add a question mark to the typescript name.
-	// This has no effect when set inside a global override; it's type specific.
-	Optional bool `json:"optional" toml:"optional" xml:"optional" yaml:"optional"`
 	// Tag is the tag name to use for the struct member(s). Default is "json".
 	Tag string `json:"tag" toml:"tag" xml:"tag" yaml:"tag"`
 	// Comment is a comment to add to the typescript interface.
 	Comment string `json:"comment" toml:"comment" xml:"comment" yaml:"comment"`
+	// Setting optional to true will add a question mark to the typescript name.
+	// This has no effect when set inside a global override; it's type specific.
+	Optional bool `json:"optional" toml:"optional" xml:"optional" yaml:"optional"`
 	// Setting KeepBadChars to true will keep bad characters in the typescript name.
 	// These include pretty much all those characters on the number keys on your keyboard.
 	KeepBadChars bool `json:"keepBadChars" toml:"keep_bad_chars" xml:"keep-bad-chars" yaml:"keepBadChars"`
@@ -70,9 +73,6 @@ type Override struct {
 	UsePkgName UsePkgName `json:"usePkgName" toml:"use_pkg_name" xml:"use-pkg-name" yaml:"usePkgName"`
 	// By default all typescript interfaces are exported. Set NoExport to true to prevent that.
 	NoExport bool `json:"noExport" toml:"no_export" xml:"no-export" yaml:"noExport"`
-	// Namer is a function that can be used to customize the typescript interface name.
-	// Use this to add a prefix, suffix or any custom name changes you wish.
-	Namer Namer `json:"-" toml:"-" xml:"-" yaml:"-"`
 }
 
 // Namer is an interface that allows external interface naming.
@@ -134,7 +134,7 @@ func (c *Config) setup() *Config {
 	c.GlobalOverrides.Name = ""
 
 	if c.DocHandler == nil {
-		c.DocHandler = goatface.NoDocs()
+		c.DocHandler = gotyface.NoDocs()
 	}
 
 	return c
